Move Bert to NLP. Tasks are moved to nlp/bert/ Refactor basic utils to modeling/
PiperOrigin-RevId: 269600561
