+++ official/bert/benchmark/bert_benchmark.py
@@ -29,9 +29,9 @@ from absl.testing import flagsaver
 import tensorflow as tf
 # pylint: enable=g-bad-import-order
 
-from official.bert import modeling
-from official.bert import run_classifier
 from official.bert.benchmark import benchmark_utils
+from official.nlp import bert_modeling as modeling
+from official.nlp.bert import run_classifier
 from official.utils.misc import distribution_utils
 
 # pylint: disable=line-too-long
