+++ official/nlp/bert/create_finetuning_data.py
@@ -24,8 +24,8 @@ from absl import app
 from absl import flags
 import tensorflow as tf
 
-from official.bert import classifier_data_lib
-from official.bert import squad_lib
+from official.nlp.bert import classifier_data_lib
+from official.nlp.bert import squad_lib
 
 FLAGS = flags.FLAGS
